fix bugs: Clang -O2 couldn't calloc() and partially passed parameters
1. When compiled with Clang -O2, the target executable would encounter
a problem that calloc() fails. This is because that in readpgm_float()
function, 'format' is declared as 'char [2]' to hold 2 characters. But
we still need 1 more place for '\0'. Clang with -O2 is affected by
this bounds overflow.

2. At least on macOS, programs compiled by either gcc or Clang can not
correctly read all the parameters. Add a colon to the 'getopt' expression
solves this problem.